Initial release of rough concepts for keepReleases and skipGitCheck
I'm finding some errors in the default code, however, for when you have a custom releases directory or custom currentPath it errors out with rsync error code 12.  It's because the directory isn't getting created or something initially, but then if you run it a second time it works... very strange.  I'm guessing it has to do with running Promises.all instead of trickling the promises since they resolve at different times.